+++ pkg/payments/estimated_price.go
@@ -1,6 +1,7 @@
 package payments
 
 import (
+	"errors"
 	"fmt"
 	"net/url"
 
@@ -32,6 +33,33 @@ func EstimatedPrice(amount float64, currencyFrom, currencyTo string) (*Estimate,
 		RouteName: "estimate",
 		Into:      &e,
 		Values:    u,
+	}
+	err := core.HTTPSend(par)
+	if err != nil {
+		return nil, err
+	}
+	return e, nil
+}
+
+// LatestEstimate holds info about the last price estimation.
+type LatestEstimate struct {
+	PaymentID      string  `json:"id"`
+	TokenID        string  `json:"token_id"`
+	PayAmount      float64 `json:"pay_amount"`
+	ExpirationDate string  `json:"expiration_estimate_date"`
+}
+
+// RefreshEstimatedPrice gets the current estimate on the payment and update
+// the current estimate.
+func RefreshEstimatedPrice(paymentID string) (*LatestEstimate, error) {
+	if paymentID == "" {
+		return nil, errors.New("missing paymentID")
+	}
+	e := &LatestEstimate{}
+	par := &core.SendParams{
+		RouteName: "last-estimate",
+		Into:      &e,
+		Path:      paymentID + "/update-merchant-estimate",
 	}
 	err := core.HTTPSend(par)
 	if err != nil {